Refactor broker ACL code to allow EE-specific topics
