+++ forge/comms/commsClient.js
@@ -8,50 +8,56 @@ class CommsClient extends EventEmitter {
     }
 
     async init () {
-        const brokerConfig = {
-            clientId: 'forge_platform',
-            username: 'forge_platform',
-            password: await this.app.settings.get('commsToken'),
-            reconnectPeriod: 5000
-        }
-        this.client = mqtt.connect(this.app.config.broker.url, brokerConfig)
-        this.client.on('connect', () => {
-            this.app.log.info('Connected to comms broker')
-        })
-        this.client.on('reconnect', () => {
-            this.app.log.info('Reconnecting to comms broker')
-        })
-        this.client.on('error', (err) => {
-            this.app.log.info(`Connection error to comms broker: ${err.toString()}`)
-        })
-        this.client.on('message', (topic, message) => {
-            const topicParts = topic.split('/')
-            const ownerType = topicParts[3]
-            const ownerId = topicParts[4]
-            if (ownerType === 'p') {
-                this.emit('status/project', {
-                    id: ownerId,
-                    status: message.toString()
-                })
-            } else if (ownerType === 'd') {
-                this.emit('status/device', {
-                    id: ownerId,
-                    status: message.toString()
-                })
+        if (this.app.config.broker.url !== ':test:') {
+            const brokerConfig = {
+                clientId: 'forge_platform',
+                username: 'forge_platform',
+                password: await this.app.settings.get('commsToken'),
+                reconnectPeriod: 5000
             }
-        })
-        this.client.subscribe([
-            'ff/v1/+/p/+/status',
-            'ff/v1/+/d/+/status'
-        ])
+            this.client = mqtt.connect(this.app.config.broker.url, brokerConfig)
+            this.client.on('connect', () => {
+                this.app.log.info('Connected to comms broker')
+            })
+            this.client.on('reconnect', () => {
+                this.app.log.info('Reconnecting to comms broker')
+            })
+            this.client.on('error', (err) => {
+                this.app.log.info(`Connection error to comms broker: ${err.toString()}`)
+            })
+            this.client.on('message', (topic, message) => {
+                const topicParts = topic.split('/')
+                const ownerType = topicParts[3]
+                const ownerId = topicParts[4]
+                if (ownerType === 'p') {
+                    this.emit('status/project', {
+                        id: ownerId,
+                        status: message.toString()
+                    })
+                } else if (ownerType === 'd') {
+                    this.emit('status/device', {
+                        id: ownerId,
+                        status: message.toString()
+                    })
+                }
+            })
+            this.client.subscribe([
+                'ff/v1/+/p/+/status',
+                'ff/v1/+/d/+/status'
+            ])
+        }
     }
 
     publish (topic, payload) {
-        this.client.publish(topic, payload)
+        if (this.client) {
+            this.client.publish(topic, payload)
+        }
     }
 
     async disconnect () {
-        this.client.end()
+        if (this.client) {
+            this.client.end()
+        }
     }
 }
